requirements.txt
# Core Dependencies
numpy>=1.24.0
python-dotenv>=1.0.0
# hermes-agent>=1.0.0
PyYAML>=6.0

# RL Dependencies
torch>=2.0.0

# RAG Knowledge Base Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0

# LLM Provider Dependencies
openai>=1.0.0
anthropic>=0.20.0

# Development & Testing (Optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# pip install git+https://github.com/NousResearch/hermes-agent.git